requirements.txt
fastapi~=0.111.0
fastapi-utilities~=0.2.0
hypercorn~=0.17.3
GitPython~=3.1.43
